Adding platform PALs so we dont introduce breaking changes
