+++ src/System.Device.Gpio/System/Device/Gpio/Drivers/HummingBoardDriver.cs
@@ -2,13 +2,42 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System.Threading;
+using System.Threading.Tasks;
+
 namespace System.Device.Gpio.Drivers
 {
     /// <summary>
     /// A GPIO driver for the HummingBoard.
     /// </summary>
-    public class HummingBoardDriver : Windows10Driver
+    public class HummingBoardDriver : GpioDriver
     {
+        /// <summary>
+        /// We can't have HummingBoardDriver inherit from Windows10Driver since that would cause issues
+        /// when we decide to implement the driver on Linux. Having an internal driver instead that we
+        /// will forward the calls for now.
+        /// </summary>
+        private Windows10Driver _internalDriver;
+
+        public HummingBoardDriver()
+        {
+            if (Environment.OSVersion.Platform != PlatformID.Win32NT)
+            {
+                throw new PlatformNotSupportedException($"{GetType().Name} is only supported on Windows.");
+            }
+
+            _internalDriver = new Windows10Driver();
+        }
+
+        /// <inheritdoc/>
+        protected internal override int PinCount => _internalDriver.PinCount;
+
+        /// <inheritdoc/>
+        protected internal override void AddCallbackForPinValueChangedEvent(int pinNumber, PinEventTypes eventTypes, PinChangeEventHandler callback) => _internalDriver.AddCallbackForPinValueChangedEvent(pinNumber, eventTypes, callback);
+
+        /// <inheritdoc/>
+        protected internal override void ClosePin(int pinNumber) => _internalDriver.ClosePin(pinNumber);
+
         /// <summary>
         /// Converts a board pin number to the driver's logical numbering scheme.
         /// </summary>
@@ -38,6 +67,38 @@ namespace System.Device.Gpio.Drivers
         protected internal override PinMode GetPinMode(int pinNumber)
         {
             throw new NotImplementedException();
+        }
+
+        /// <inheritdoc/>
+        protected internal override bool IsPinModeSupported(int pinNumber, PinMode mode) => _internalDriver.IsPinModeSupported(pinNumber, mode);
+
+        /// <inheritdoc/>
+        protected internal override void OpenPin(int pinNumber) => _internalDriver.OpenPin(pinNumber);
+
+        /// <inheritdoc/>
+        protected internal override PinValue Read(int pinNumber) => _internalDriver.Read(pinNumber);
+
+        /// <inheritdoc/>
+        protected internal override void RemoveCallbackForPinValueChangedEvent(int pinNumber, PinChangeEventHandler callback) => _internalDriver.RemoveCallbackForPinValueChangedEvent(pinNumber, callback);
+
+        /// <inheritdoc/>
+        protected internal override void SetPinMode(int pinNumber, PinMode mode) => _internalDriver.SetPinMode(pinNumber, mode);
+
+        /// <inheritdoc/>
+        protected internal override WaitForEventResult WaitForEvent(int pinNumber, PinEventTypes eventTypes, CancellationToken cancellationToken) => _internalDriver.WaitForEvent(pinNumber, eventTypes, cancellationToken);
+
+        /// <inheritdoc/>
+        protected internal override ValueTask<WaitForEventResult> WaitForEventAsync(int pinNumber, PinEventTypes eventTypes, CancellationToken cancellationToken) => _internalDriver.WaitForEventAsync(pinNumber, eventTypes, cancellationToken);
+
+        /// <inheritdoc/>
+        protected internal override void Write(int pinNumber, PinValue value) => _internalDriver.Write(pinNumber, value);
+
+        /// <inheritdoc/>
+        protected override void Dispose(bool disposing)
+        {
+            _internalDriver?.Dispose();
+            _internalDriver = null;
+            base.Dispose(disposing);
         }
     }
 }
