+++ src/System.Device.Gpio/System/Device/Gpio/Drivers/RaspberryPiRegisterDriver.cs
@@ -1,0 +1,78 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Device.Gpio.Drivers;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace System.Device.Gpio
+{
+    internal class RaspberryPiRegisterDriver : GpioDriver
+    {
+        private GpioDriver _internalDriver;
+
+        public RaspberryPiRegisterDriver()
+        {
+            _internalDriver = new Windows10Driver();
+        }
+
+        protected internal virtual ulong ClearRegister
+        {
+            get { throw new PlatformNotSupportedException(); }
+            set { throw new PlatformNotSupportedException(); }
+        }
+
+        protected internal virtual ulong SetRegister
+        {
+            get { throw new PlatformNotSupportedException(); }
+            set { throw new PlatformNotSupportedException(); }
+        }
+
+        /// <inheritdoc/>
+        protected internal override int PinCount => _internalDriver.PinCount;
+
+        /// <inheritdoc/>
+        protected internal override void AddCallbackForPinValueChangedEvent(int pinNumber, PinEventTypes eventTypes, PinChangeEventHandler callback) => _internalDriver.AddCallbackForPinValueChangedEvent(pinNumber, eventTypes, callback);
+
+        /// <inheritdoc/>
+        protected internal override void ClosePin(int pinNumber) => _internalDriver.ClosePin(pinNumber);
+
+        /// <inheritdoc/>
+        protected internal override int ConvertPinNumberToLogicalNumberingScheme(int pinNumber) => _internalDriver.ConvertPinNumberToLogicalNumberingScheme(pinNumber);
+
+        /// <inheritdoc/>
+        protected internal override PinMode GetPinMode(int pinNumber) => _internalDriver.GetPinMode(pinNumber);
+
+        /// <inheritdoc/>
+        protected internal override bool IsPinModeSupported(int pinNumber, PinMode mode) => _internalDriver.IsPinModeSupported(pinNumber, mode);
+
+        /// <inheritdoc/>
+        protected internal override void OpenPin(int pinNumber) => _internalDriver.OpenPin(pinNumber);
+
+        /// <inheritdoc/>
+        protected internal override PinValue Read(int pinNumber) => _internalDriver.Read(pinNumber);
+
+        /// <inheritdoc/>
+        protected internal override void RemoveCallbackForPinValueChangedEvent(int pinNumber, PinChangeEventHandler callback) => _internalDriver.RemoveCallbackForPinValueChangedEvent(pinNumber, callback);
+
+        /// <inheritdoc/>
+        protected internal override void SetPinMode(int pinNumber, PinMode mode) => _internalDriver.SetPinMode(pinNumber, mode);
+
+        /// <inheritdoc/>
+        protected internal override WaitForEventResult WaitForEvent(int pinNumber, PinEventTypes eventTypes, CancellationToken cancellationToken) => _internalDriver.WaitForEvent(pinNumber, eventTypes, cancellationToken);
+
+        /// <inheritdoc/>
+        protected internal override ValueTask<WaitForEventResult> WaitForEventAsync(int pinNumber, PinEventTypes eventTypes, CancellationToken cancellationToken) => _internalDriver.WaitForEventAsync(pinNumber, eventTypes, cancellationToken);
+
+        /// <inheritdoc/>
+        protected internal override void Write(int pinNumber, PinValue value) => _internalDriver.Write(pinNumber, value);
+
+        protected override void Dispose(bool disposing)
+        {
+            _internalDriver?.Dispose();
+            _internalDriver = null;
+            base.Dispose(disposing);
+        }
+    }
+}
